flip isAssignableFrom logic
Unless I'm completely mistaken, Spring and *all* other examples online have
this backwards. We want to test if the type we produce is assignable to the
parameter. The reversed logic is not typesafe; requiring "Map or some subtype
of Map", for instance, works for parameters of type "HashMap", if that is what
we provided, but crashes at runtime for parameters of type "TreeMap" - we
shouldn't be claiming to support "TreeMap"!

The only advantage to the backwards way is that it does not match "Object". If
this is important, either we should exclude "Object", or perform exact matches
instead of checking for assignability.